backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
google-auth==2.38.0
google-genai==1.8.0
google-api-python-client==2.165.0
google-auth-httplib2==0.2.0
slowapi==0.1.9
redis==5.2.1
python-multipart==0.0.20
pytest==8.3.5
pytest-asyncio==0.25.3
